[refacto] PhotoBoard is now a container
